services/collector/pyproject.toml
[project]
name = "collector"
version = "0.1.0"
description = "Kainex multi-market data collector service"
readme = "README.md"
authors = [
    { name = "francismiko", email = "1485059943@qq.com" }
]
requires-python = ">=3.12"
dependencies = [
    "akshare",
    "ccxt",
    "yfinance",
    "pydantic",
    "pydantic-settings",
    "exchange-calendars",
    "aiosqlite>=0.22.1",
    "duckdb>=1.0.0",
    "pyarrow",
    "croniter",
    "finnhub-python",
    "baostock",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
]

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"
